// src/modules/tg-bot/constants/scenes.id.enum.ts
export enum ENUM_SCENES_ID {
    // common scenes
    START_SCENE_ID = 'START_SCENE_ID',
    ABOUT_SCENE_ID = 'ABOUT_SCENE_ID',
    HELP_SCENE_ID = 'HELP_SCENE_ID',
    HOME_SCENE_ID = 'HOME_SCENE_ID',

    // admin scenes
    ADMIN_SCENE_ID = 'ADMIN_SCENE_ID',
    ADD_ADMIN_SCENE_ID = 'ADD_ADMIN_SCENE_ID',
    DELETE_ADMIN_SCENE_ID = 'DELETE_ADMIN_SCENE_ID',

    // CHARACTER
    CREATE_CHARACTER_FORM_SCENE_ID = 'CREATE_CHARACTER_FORM_SCENE_ID',
    PROFILE_SCENE_ID = 'PROFILE_SCENE_ID',
    CHARACTER_PARAMETERS_SCENE_ID = 'CHARACTER_PARAMETERS_SCENE_ID',
    // BACKGROUND
    BACKGROUND_SCENE_ID = 'BACKGROUND_SCENE_ID',
    EDIT_NAME_SCENE_ID = 'EDIT_NAME_SCENE_ID',
    EDIT_AVATAR_SCENE_ID = 'EDIT_AVATAR_SCENE_ID',
    EDIT_CHARACTER_HISTORY_SCENE_ID = 'EDIT_CHARACTER_HISTORY_ID',
    //GRIMOIRE
    GRIMOIRE_SCENE_ID = 'GRIMOIRE_SCENE_ID',
    CREATE_SPELL_FORM_SCENE_ID = 'CREATE_SPELL_FORM_SCENE_ID',
    EDIT_MAGIC_NAME_SCENE_ID = 'EDIT_MAGIC_NAME_SCENE_ID',
    //INVENTORY
    INVENTORY_SCENE_ID = 'INVENTORY_SCENE_ID',
    EQUIPMENT_SCENE_ID = 'EQUIPMENT_SCENE_ID',

    //DEVILS
    ALL_DEVILS_SCENE_ID = 'ALL_DEVILS_SCENE_SCENE_ID',
    ALL_DEVILS_BY_RANK_SCENE_ID = 'ALL_DEVILS_BY_RANK_SCENE_ID',
    ALL_DEVILS_BY_FLOOR_SCENE_ID = 'ALL_DEVILS_BY_FLOOR_SCENE_ID',
    MY_DEVILS_SCENE_ID = 'MY_DEVILS_SCENE_SCENE_ID',
    CREATE_DEVIL_FORM_SCENE_ID = 'CREATE_DEVIL_FORM_SCENE_ID',
    //SPIRITS
    ALL_SPIRITS_SCENE_ID = 'ALL_SPIRITS_SCENE_ID',
    MY_SPIRITS_SCENE_ID = 'MY_SPIRITS_SCENE_ID',
    CREATE_SPIRIT_SCENE_ID = 'CREATE_SPIRIT_SCENE_ID',

    //MAP
    MAP_SCENE_ID = 'MAP_SCENE_ID',
    //ORGANIZATIONS
    ORGANIZATIONS_SCENE_ID = 'ORGANIZATIONS_SCENE_ID',
    // ARMY
    ARMED_FORCES_SCENE_ID = 'ARMED_FORCES_SCENE_ID',
    ARMY_REQUEST_ACCEPT_SCENE_ID = 'ARMY_REQUEST_ACCEPT_SCENE_ID',
    ARMY_REQUEST_REJECT_SCENE_ID = 'ARMY_REQUEST_REJECT_SCENE_ID',
    CREATE_SQUAD_SCENE_ID = 'CREATE_SQUAD_SCENE_ID',
    //MONEY
    WALLET_SCENE_ID = 'WALLET_SCENE_ID',
    WORK_SCENE_ID = 'WORK_SCENE_ID',
    // MINES
    MINES_SCENE_ID = 'MINES_SCENE_ID',

    // MAGIC PARLAMENT
    MAGIC_PARLAMENT_SCENE_ID = 'MAGIC_PARLAMENT_SCENE_ID',
    REQUEST_TO_PARLAMENT_SCENE_ID = 'REQUEST_TO_PARLAMENT_SCENE_ID',
    CREATE_REQUEST_TO_PARLAMENT_SCENE_ID = 'CREATE_REQUEST_TO_PARLAMENT_SCENE_ID',
}